Fold remaining tab methods from window.py into WindowTabsMixin
_page_for_child, _select_tab_relative, _move_tab_relative,
_close_active_tab_or_pane, _focus_most_recent_tab(_or_open_new),
_cycle_connection_tabs_or_open, _focus_terminal_widget,
_get_active_terminal_widget, on_tab_selected, _update_tab_titles, and the
_is_terminal_widget helper now live with the rest of the tab lifecycle in
window_tabs.py. Pure code move.
